+++ app/scripts/controllers/transfer.js
@@ -23,19 +23,48 @@ angular.module('swaptricksApp')
   });
   // Trade Entyties
   $scope.userTransfers;
+  $scope.intoTransfers;
+  $scope.productOne;
+  $scope.productTwo;
 
   //Load user transfers
-  $scope.chargeTransfers = function() {
-    $http.get(`http://api.swapingzone.com:3000/transfers?token=${localStorage.token}`)
+  $scope.chargeTransfers = function(pDeal) {
+    if(!pDeal)
+    var url = `http://api.swapingzone.com:3000/transfers?token=${localStorage.token}`;
+    else
+    var url = `http://api.swapingzone.com:3000/transfers?token=${localStorage.token}&deal=true`;
+    $http.get(url)
     .then(function successCallback(responce){
       console.log(responce);
+      if(!pDeal)
       $scope.userTransfers = responce.data;
+      else
+      $scope.intoTransfers = responce.data;
+    }, function errorCallback(responce){
+      console.log(responce);
+    });
+  }
+  //Deal: Close the deal!
+  $scope.closeDeal = function(trade){
+    //make the Call
+    var data = {"product_req_id": trade.product_req_id,
+    "product_offer_id": trade.product_offer_id,"active": false, "state": "close",
+    "user_id": trade.user_id, "product_offer_name": trade.product_offer_name,
+    "product_req_name": trade.product_req_name, "to_whom": trade.to_whom};
+    console.log(trade.id);
+    trade.active = false;
+    trade.state = "close";
+    console.log(trade);
+    $http.put(`http://api.swapingzone.com:3000/transfers/${trade.id}?token=${localStorage.token}`, data)
+    .then(function successCallback(responce){
+      console.log(responce);
+      $scope.chargeTransfers(true);
     }, function errorCallback(responce){
       console.log(responce);
     });
   }
 
-  //TRADE!
+  //TRADE try a new trade!
   $scope.trade = function(pSelected) {
     if (pSelected == undefined || pSelected == null) {
       alert("select your product before trade it");
@@ -46,7 +75,8 @@ angular.module('swaptricksApp')
       //Proces  the trade
       var data = {"product_req_id": $scope.productSelected.id,
       "product_offer_id": pSelected.id,"active": true, "state": "?",
-      "user_id": localStorage.type};
+      "user_id": localStorage.type, "product_offer_name":"_",
+      "product_req_name": "_", "to_whom": $scope.productSelected.user_id};
       $scope.createTransfer(data);
     }
     console.log("nope");
@@ -66,16 +96,20 @@ angular.module('swaptricksApp')
 
   // Delete transfer
   $scope.deleteTransfer = function(pTransfer){
-    var url = `http://api.swapingzone.com:3000/transfers/${pTransfer.id}?token=${localStorage.token}`;
-    $http.delete(url).then(function successCallback(responce){
-      alert("Deal Broken");
-      console.log(responce);
-      window.location = "http://localhost:9000/#/dashboard#";
-    }, function errorCallback(responce){
-      alert("Upps... someting went wrong, try again");
-    });
+    var res = confirm("Are you sure?");
+    if (res) {
+      var url = `http://api.swapingzone.com:3000/transfers/${pTransfer.id}?token=${localStorage.token}`;
+      $http.delete(url).then(function successCallback(responce){
+        alert("Deal Broken");
+        console.log(responce);
+        window.location = "http://localhost:9000/#/dashboard#";
+      }, function errorCallback(responce){
+        alert("Upps... someting went wrong, try again");
+      });
+    }
   }
 
   //Execute loadTransfer function
-  $scope.chargeTransfers();
+  $scope.chargeTransfers(true);
+  $scope.chargeTransfers(false);
 });
